read the description
hashmap is in memory and is generated every time the program starts for now.
the terminal will print the query results as Docids and similarity with query.
Queries are instant but that might change when I save the hashmap on a file. Need to search lucene index with a  docid query to extract the document info